Post tell don't ask
Exposes the principle with a catch: if people really understand it, they
should point out that the discount behavior should be on the product and
not on the discount itself.

Still, I tried to make the most compelling examples I could, considering
the shortest changes from previous examples. For those who can see
throught them, there are valuable lessons on how they can create their
MVPs and progress through the evolution of their product. Hopefully,
using an e-commerce example is better than a blog and still common
enough so web developers can relate to it.

For the next post, I have serious doubts on how to proceed. One thing I
can surely do is a refactor, but I've though on many things that I can
related to this and the previous post:

- How these posts help you unit test your classes. I think that exposing
  the difficulties of unit testing as of OO practice and lack of
  knowledge by people would be a great thing.
- I could also express how the "Discount" classes are a Strategy pattern
  and how we can related the Null Object pattern with them to make up
  for clearer code implementations.
- I could create a post with the same from previous posts, violating
  both "principles", so people can see through the benefist and judge
  the tradeoffs for themselves.

If anybody is reading this, feedback is greatly appreciated.