A post on how to become a better developer
I doubt many people read commit messages of projects, specially blogs as
I do. This is such a rarity that I scripted my way out of it so whenever
I clone a project I search for a (bacon) ratio of >.5 between commit messages
and the diff stat. It is rare, but when that happens I usually spend
days reading them. **So satisfying**. Every time!

The idea behind the post is to make people grok Unix. Not that I do. But
I must say, I learn a trick every day. Just using awk.

    $ cat many-duplicated-lines.csv | awk '!$dupes[$2]++'

Made bit.lys all over the links so I am able to ignore the tracking it
gives me to know if people really follow links I find interesting - and
make them part of the argument - or not.

If you are reading this and haven't watched "Jiro Dreams of Sushi", go
do it. I will leave the suggestion of "Time of Eve" to a next post.

See ya.

PS: I fail my own *bacon ratio*, on purpose. Humpf.