Keep only speeches I am willing to present
Rather than keeping every talk I've ever given, a document with the
subjects I am interested in (and have material) currently is more
interesting to share with other people.

Making this file a history of speeches I gave is good, so I can have the
best of both worlds with Git: the history and a sane file which can
acutally be read.