Adds Object Oriented talk description
A good idea in the future should be storing all the files I use to
create a talk in this repository as well. Feeling brilliant and stupid
at the same time here.

One thing I don't know yet is if it is usefull to put my handwritten
notes here as well, as of now, I can't find of any good reasons not to.
At least I could own all the data and find them in the same place,
instead of recurring to Dropbox, Evernote or physical Moleskines.

----

Mental note, purchase another batch of Moleskines. I am currently using
10 large plain soft notebooks a year, while I should be using more I was
under the impression this last batch would take 2-3 years to use. Nice.