Supports "Revisiting OO" talk with code
I should probably write a post instead of giving a talk about it, but
who cares? I had to talk about something and OO was too high on my
mind/reading/talking at the time - and still is.

I remember [giving a talk][1] I hold no sentiment to. Not that I've lied
or dislike the content, but everything can be improved: examples,
real-case scenarios, rhythm, the conclusion and the questioning. I love
to give talks where every slide has a good book behind it or a twist,
this had none - only a soup of letters so people can perform well on
their next lame interviews.

I was, also, too pretentious on using the DOM as an OO example of
composition. Pretentious with my capability of explaining how it works in
30s in a way people could understand. I definetly need more than 30s
which I may be able to achieve when I reorder stuff around and build up
everything towards it. If people grok DOM design, they will grok OO -
fuck, I did (grok DOM).

Note to self: 700ml of Bushmills White awakes you.
Note to self 2: Please, put yourself together some day and order that
damn talk - it can be made good.

[1]: https://bit.ly/pascutti-oo