Post "PHP and Exceptions"
There is a lot of texts out there saying "Exceptions are harmful" - pun
intended. They are, when being "Go to"s in disguise, but they can also
be useful.

My idea is to come out with one example where exceptions are good and
unnecessary at the same time - a validation. It may, as well, be a
sequence to posts about OO some day.