:scissors: Favors Jekyll plugins
Rather than implementing everything myself as before, uses a plugin for
Feed and Sitemap generation. Allows to remove files we don;t need to
maintain.

A *redirect plugin* is also used to redirect `/blog/` url to `/posts`
without fucking with people who might've linked to this URL  (probably
only me). This will also be useful to fix the permalinks for posts that
are currently considering their category and best would be just to have
date and slug.