Post about test automation
I've spoken about test automation numerous times, this is my most recent
attempt.

I am planning on having a post for each talk I would like to give again
with material people can use to continue their learning on that area or
the references I've used. This is the first post on that "planned
series" which:

* Create a page "Talks" with posts and a brief description of talks
* Allow people to evaluate them in some way that conserves the data with
me and not a 3rd party site
* Allow people to link content to a permanent URL I manage