Broken links action fails if any is found
Updating the action provided a breaking change that makes sense
absorbing. The action step now fails if a broken link is found. It
is more adequate behavior to something akin to a CI check.

Rather than specifying a patch version, focus just on the major version
to allow the action to use the latest compatible version - since the
author respects semantic versioning properly.

Configuration is controlled in a TOML file, i grabbed the full
configuration example from the tool's documentation.